@@ -19,17 +19,18 @@
     "@testing-library/jest-dom": "^4.2.4",
     "@testing-library/react": "^9.5.0",
     "@testing-library/user-event": "^7.2.1",
-    "axios": "^0.19.2",
-    "cordova-android": "^8.1.0",
-    "cordova-ios": "^6.1.0",
+    "axios": "^0.21.1",
+    "compare-func": "^2.0.0",
+    "cordova-android": "^9.1.0",
+    "cordova-ios": "^6.2.0",
     "cordova-plugin-camera": "^4.1.0",
     "cordova-plugin-device": "^2.0.3",
     "cordova-plugin-statusbar": "^2.4.3",
     "cordova-res": "^0.15.1",
-    "react": "^16.13.1",
-    "react-cropper": "^1.3.0",
-    "react-dom": "^16.13.1",
-    "react-scripts": "3.4.1",
+    "react": "^16.14.0",
+    "react-cropper": "^2.1.7",
+    "react-dom": "^16.14.0",
+    "react-scripts": "^4.0.3",
     "rimraf": "^3.0.2"
   },
   "browserslist": {
@@ -50,6 +51,7 @@
   "author": "Artsdatabanken",
   "license": "MIT",
   "devDependencies": {
+    "cordova-plugin-splashscreen": "^6.0.0",
     "cordova-plugin-whitelist": "^1.3.4"
   },
   "cordova": {
