@@ -1,10 +1,7 @@
 import React from "react";
 import "../App.css";
-import ReportButton from "./ReportButton";
-import Button from "@material-ui/core/Button";
-import WarningIcon from "@material-ui/icons/Warning";
 
-function IdResult({ result, openDialog,croppedImages }) {
+function IdResult({ result, openResult, croppedImages }) {
   const percentage = result.probability * 100;
   const strokes = percentage.toString() + " " + (100 - percentage).toString();
   const colors = [
@@ -30,9 +27,12 @@ function IdResult({ result, openDialog,croppedImages }) {
   };
 
   let color = getColor(percentage);
+  const openResultModal = () => {
+    openResult(result);
+  }
 
   return (
-    <div className="resultRow">
+    <div className="resultRow" onClick={openResultModal}>
       <div className="resultDonut">
         <svg width="100%" height="100%" viewBox="0 0 42 42">
           <circle
@@ -41,7 +41,7 @@ function IdResult({ result, openDialog,croppedImages }) {
             r="15.91549430918954"
             fill="transparent"
             stroke="#d2d3d4"
-            strokeWidth="6"
+            strokeWidth="7"
           ></circle>
 
           <circle
@@ -50,11 +50,12 @@ function IdResult({ result, openDialog,croppedImages }) {
             r="15.91549430918954"
             fill="transparent"
             stroke={color}
-            strokeWidth="6"
+            strokeWidth="7"
             strokeDasharray={strokes}
-            strokeDashoffset="0"
+            strokeDashoffset="20"
           ></circle>
         </svg>
+        <div className="percentage">{Math.round(percentage)}%</div>
       </div>
 
       <div className="resultLabels">
@@ -70,46 +71,21 @@ function IdResult({ result, openDialog,croppedImages }) {
             result.taxon.vernacularName.slice(1)}
         </div>
         <div className="scientific">{result.taxon.name}</div>
-        <div className="percentage">
-          <span className="number">({Math.round(percentage)} % treff)</span>
-        </div>
-        <div className="group">
-          {result.taxon.groupName}{" "}
-          {result.taxon.groupName === "Sopper" && (
-            <span className="danger">
-              <WarningIcon /> ALDRI SPIS NOE BASERT PÅ ARTSORAKELETS SVAR
-            </span>
-          )}
-        </div>
-        <div className="actions">
-          <a
-            href={result.taxon.infoUrl}
-            target={"_blank"}
-            onClick={
-              result.taxon.scientificNameID
-                ? () => {
-                    return;
-                  }
-                : (e) => {
-                    e.preventDefault();
-                  }
-            }
-            rel="noopener noreferrer"
-          >
-            <Button
-              style={{
-                fontSize: "9px",
-                lineHeight: "10px",
-              }}
-              variant="contained"
-              color="primary"
-            >
-              Les mer
-            </Button>
-          </a>
+        <div className="group">{result.taxon.groupName}</div>
+        {result.taxon.groupName === "Sopper" && (
+          <div className="danger">
+            ALDRI SPIS NOE PGA APPEN
+          </div>
+        )}
+      </div>
 
-          <ReportButton reportResult={result} croppedImages={croppedImages}/>
-        </div>
+      <div className="chevron-right">
+        <svg viewBox="0 0 24 24">
+          <path
+            fill="currentColor"
+            d="M8.59,16.58L13.17,12L8.59,7.41L10,6L16,12L10,18L8.59,16.58Z"
+          />
+        </svg>
       </div>
     </div>
   );
